Created CalendarsI interface in calfacade/ifs - it's the Calendars section of Calintf
Created separate Calendars component in calcore which is called by CalintImpl to do the work. It implements CalendarsI.

Added new code to Calendars to check access of calendar (sub)tree we are trying to fetch and truncate it at inaccessible entries.

ALso partially fix up calendar deletions. Still an error trying to delet parent but seems to be a stylesheet problem.