Two sets of changes - bwconfig and session stuff
BwConfig: very out of date. 
Started on changes to bring it up to date with all the property changes that took place. 
Still broken but does build. More to follow.

Sessions - packaging the suite as one ear file revealed a number of problems with tracking session counts.
Though only intended as a guide to usage they are useful and the single static field became
an accumulation of all applications use. Switched to a table based approach using the 
app name as a key.