Deletions mostly working.
New svci method markDeleted which marks events as deleted and moves them into the owners trash

For events to which the user has no access, e.g. public events, an annotation is added which is 
marked as deleted and is also stored in the trash

Further work:
Need to handle recurrence instances.
Need a way to really delete events
Need to ensure empty trash doesn't remove the annotations.