More uportal integration
When running as a portlet will now suffix the approot with "_" + protletPlatform.

Allows us to have a separate set of stylesheets for portlet mode as against stand-alone mode.

Committed a copy of the current set of standalone stylesheets ready for customization.

Also made more run-time properties available via the options object.