Yet more changes for portlet and calendar suites.
These changes add the personal client to the partially portletised set of applications.

In addition changes have been made to how xslt is located.

Starting with the appRoot - if running under a portal the approot has "." + portalPlatform appended

For example we might have an approot of

http://www.mysite.edu/xslthome/cal

If it's running uder uportal the root becomes   

http://www.mysite.edu/xslthome/cal.uPortal2

In addition, we append the calendar suite name for public clients.
So if the calendar suite is MainCampus we have

http://www.mysite.edu/xslthome/cal.MainCampus

or for the portlet
http://www.mysite.edu/xslthome/cal.uPortal2.MainCampus