Significant changes to handling of calendars...
Db entities are cached (and shared) so the children property of the db object must always refelct th actual 
db structure.

An approach was taken of cloning the objects and returning a tree to the user which reflected their access.

This had some problems. The new approach is to create a getCalendars(cal) method which returns the 
children of the given calendar to which the user has access.

This does have some implications for developers of applications and clients.

Calls to the calendar getChildren() method will return ALL children of the calendar WITHOUT access checks.