Two major sets of changes going on here
Continuation of adding language properties to appropriate strings. Location is doen with this set.
Sponsors are probably broken but will be fixed soon.

Other major change is to merge admin adn webclient code into the common code. 
The current set of changes address events and locations - sponsors still need to be done and there
are still problems with events.