Change schema to avoid repeated fetch of owner/creator with entities.
Owner/creator is now stored in the db as a principal hierarchy path, a string value. The system then retrieves and caches the actual db entries as they are needed.

This should result in less jvm churn. It also solves a number of problems in allowing groups to own entities. The move to a principal object is not complete but should be relatively simple.

Almost no visible changes as a result of this change. The only exception is where owners/creators appear in the generated xml. They are now a full principal path rather than just the account.