Allow us to track changes made to a meeting:
Slight refactoring to move event specific methods into a common base class.

Change request parameters so we're not updating the event object in the form directly from the request. That is: no more event.xxx

As we go through the request parameters update an instance of ChangeTable. This can be used to drive the scheduling