Some changes that didn't make it through for the last update.
In addition to the below changes to try to get a stack trace out of jsp when rendering fails.


Further changes to the build. Attempt to catch changed libraries and force a rebuild.

Also use outofdate to suppress running of apt when sources have not changed.

Minor but widespread change to remove debug parameter from a number of classes. For those debugging now depends on log4j level.