(Partial updates as full update failed with some conflict)
Major refactoring of code to move all jar building into separate projects outside of the main bedework project which itself becomes an assembly project - that is it builds deployable components from the jars and other information.

This change is to facilitate the move to jboss 7. 

Further changes are needed but probably no more major reorganization.