Some missing dependencies in the build
Remove the "doweSynch" flag. Synch engine now required. Code removed.

Add methods and service bean actions to check external subscriptions and (re)subscribe if necessary

Tidied up synchws schema a little

Immediately after a dump or restore (not in time but in actions) click on the checkExternalSubs button to have external subscriptions checked.

This will cause subscriptions to be (re)subscribed. If this is an upgrade from an earlier system this is the way to get all those external subscriptions working in the new world.