Almost completely build changes - mostly elimination of the per-project build.properties file.
There is now a single properties file in the config which determines all the project versions - with the exception of deployutil.

This update also includes a new deployment task to collect jar references and license information. A small utility will generate a readable notice and references file for the quickstart and/or the site.