Add a saveData and deployData target.
The saveData target will save all of the necessary bedework data, the hypersonic directory and the solr directory and zip it all up inside the bedework/resources direvctory ready for commit.

The deployData target will unzip the data and replace the 3 directories inside the jboss data directory.

Before doing a saveData you should reindex and clean up the index directories to remove the old indexes.